Simplify --yes interfaces for post-alpha CLI enhancements (#53)
* Simplify --yes interfaces for post-alpha CLI enhancements

For the prompted (i.e. non-yes) interfaces, our philosophy is to prompt the user for any/all required information. For example, they run `alwaysai app configure` but they're not logged in, we simply prompt them for their credentials. For the non-prompted "yes" interfaces, this PR makes the commands fail if they're not run in order. For example, if the user runs `alwaysai app configure --yes` but they're not logged in, the command will error out with `Error: Authentication required. Please run "alwaysai user login".`

* Remove "required" restriction on path in `app configure --yes`